deleted oracle aggregator
