+++ contracts/dapp_interfaces/chainlink/IGasPriceOracle.sol
@@ -5,4 +5,6 @@ import {IERC20} from "@gelatonetwork/core/contracts/external/IERC20.sol";
 
 interface IGasPriceOracle {
     function latestAnswer() external view returns (int256);
+
+    function decimals() external view returns (uint256);
 }
