+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "omen",
+  "name": "gelato-omen",
   "version": "1.1.0",
   "engines": {
     "node": "14.15.1"
@@ -23,11 +23,10 @@
     "gnosis",
     "omen.eth"
   ],
-  "author": "",
+  "author": "hilmarx & superarius",
   "license": "ISC",
   "dependencies": {},
   "devDependencies": {
-    "@chainlink/contracts": "0.0.11",
     "@codechecks/client": "0.1.10",
     "@gelatonetwork/core": "1.5.1",
     "@gnosis.pm/conditional-tokens-contracts": "^1.0.1",
